requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
Werkzeug==0.9.6
itsdangerous==0.24
nltk==3.0.0
pytesseract==0.1.5
requests==2.5.1
wsgiref==0.1.2
